+++ contracts/mocks/Ramm/RammMockMCR.sol
@@ -37,11 +37,11 @@ contract RammMockMCR is IMCR {
     revert("Unsupported");
   }
 
-  function mcr() external override pure returns (uint112) {
+  function mcr() external override pure returns (uint80) {
     revert("Unsupported");
   }
 
-  function desiredMCR() external override pure returns (uint112) {
+  function desiredMCR() external override pure returns (uint80) {
     revert("Unsupported");
   }
 
@@ -53,11 +53,11 @@ contract RammMockMCR is IMCR {
     revert("Unsupported");
   }
 
-  function maxMCRIncrement() external pure returns (uint24) {
+  function maxMCRIncrement() external pure returns (uint16) {
     revert("Unsupported");
   }
 
-  function minUpdateTime() external pure returns (uint24) {
+  function minUpdateTime() external pure returns (uint16) {
     revert("Unsupported");
   }
 }
